rename home/Home, add stub frontend jest test
