+++ src/frontend/Router.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { browserHistory, Router, Route, IndexRoute } from 'react-router';
 
 /* Pages */
-import Home from './pages/home/Home';
+import Home from './pages/Home';
 import Layout from './pages/Layout';
 import NotFound from './pages/NotFound';
